Wire up new blocks cache config
