+++ pkg/storage/stores/shipper/bloomshipper/config/config.go
@@ -7,14 +7,15 @@ import (
 	"strings"
 	"time"
 
+	"github.com/grafana/dskit/flagext"
 	"github.com/grafana/loki/pkg/storage/chunk/cache"
 )
 
 type Config struct {
-	WorkingDirectory       string                    `yaml:"working_directory"`
-	BlocksDownloadingQueue DownloadingQueueConfig    `yaml:"blocks_downloading_queue"`
-	BlocksCache            cache.EmbeddedCacheConfig `yaml:"blocks_cache"`
-	MetasCache             cache.Config              `yaml:"metas_cache"`
+	WorkingDirectory       string                 `yaml:"working_directory"`
+	BlocksDownloadingQueue DownloadingQueueConfig `yaml:"blocks_downloading_queue"`
+	BlocksCache            BlocksCacheConfig      `yaml:"blocks_cache"`
+	MetasCache             cache.Config           `yaml:"metas_cache"`
 }
 
 type DownloadingQueueConfig struct {
@@ -37,6 +38,42 @@ func (c *Config) RegisterFlagsWithPrefix(prefix string, f *flag.FlagSet) {
 func (c *Config) Validate() error {
 	if strings.TrimSpace(c.WorkingDirectory) == "" {
 		return errors.New("working directory must be specified")
+	}
+	return nil
+}
+
+// BlocksCacheConfig represents in-process embedded cache config.
+type BlocksCacheConfig struct {
+	SoftLimit flagext.Bytes `yaml:"soft_limit"`
+	HardLimit flagext.Bytes `yaml:"hard_limit"`
+	TTL       time.Duration `yaml:"ttl"`
+
+	// PurgeInterval tell how often should we remove keys that are expired.
+	// by default it takes `defaultPurgeInterval`
+	PurgeInterval time.Duration `yaml:"-"`
+}
+
+func (cfg *BlocksCacheConfig) RegisterFlagsWithPrefix(prefix, description string, f *flag.FlagSet) {
+	cfg.RegisterFlagsWithPrefixAndDefaults(prefix, description, f, time.Hour)
+}
+
+func (cfg *BlocksCacheConfig) RegisterFlagsWithPrefixAndDefaults(prefix, description string, f *flag.FlagSet, defaultTTL time.Duration) {
+	cfg.SoftLimit.Set("32GiB")
+	f.Var(&cfg.SoftLimit, prefix+"soft-limit", description+"Soft limit of the cache in bytes. Exceeding this limit will trigger evictions of least recently used items in the background.")
+	cfg.HardLimit.Set("64GiB")
+	f.Var(&cfg.HardLimit, prefix+"hard-limit", description+"Hard limit of the cache in bytes. Exceeding this limit will block execution until soft limit is deceeded.")
+	f.DurationVar(&cfg.TTL, prefix+"ttl", defaultTTL, description+"The time to live for items in the cache before they get purged.")
+}
+
+func (cfg *BlocksCacheConfig) Validate() error {
+	if cfg.TTL == 0 {
+		return errors.New("blocks cache ttl must not be 0")
+	}
+	if cfg.SoftLimit == 0 {
+		return errors.New("blocks cache soft_limit must not be 0")
+	}
+	if cfg.SoftLimit > cfg.HardLimit {
+		return errors.New("blocks cache soft_limit must not be greater than hard_limit")
 	}
 	return nil
 }
